other_services/service_manager/requirements.txt
gmqtt==0.6.11
python-dotenv==1.0.0
colorlog==6.7.0